tensorflow2_fastapi/requirements.txt
fastapi
uvicorn
python-multipart
pillow